Supervisor-RC-Edits-to-interview-questions.md
The changes are in line and comments are included in this PR comment. These came from an annotated .docx document that the supervisor emailed to Eriol directly. 

Question topic
3. What does 'user' (or beneficiary) engagement look like to you and your organisation (could be different)
Supervisor comments: This feels more like an organisational question – have moved to earlier … Also is this a question about who their key users/beneficiaries are and identifying how they are defined by the organisation?
Eriol's response: To some extent yes, I think this question will allow me to probe into how the individual and the organisation both describe users/beneficiaries with the intention that the question framing of 'engagements' will allow for more exploration into how the users/beneficiaries engage with the individual/org rather than generalised characteristics that could be assumptions and/or flawed.

Question topic
What does the term 'innovation' [in relation to technology <- possibly delete ref to technology] mean to you within your organisational context? 
Supervisor comments: The term 'technology' could be leading here. Consider removing to see what other kinds of responses present themselves when this question is asked.
Eriol's response: Agreed, even though I want to focus on technology it's important to gather the other information and topics around technology that are part of the individual and organisational understanding of innovation (in relation to technology and more specifically, open-source software)

Question topic
Do all or most of the people working in your organisation agree on the definitions and aims of innovation? 
Supervisor comments: This could be a difficult question to ask and answer and might need to be circled back on or one that you gather information on as you ask other questions.
Eriol's response: Could start the actual ecosystem mapping part - other orgs terms and processes might be mentioned.

Question topic
How would you describe how decisions are made regarding innovation internally and with partners or stakeholders? 
Eriol's comments: (I am) Researcher is looking for hierarchical systems, adhocratic, sociocratic or holocractic mentions)
Supervisor comments: Sure, these are quite obscure terms though so you might need to do mapping here based off the source paper information.